pyproject.toml
[project]
name = "permit-copilot"
version = "0.1.0"
description = "AI copilot for building permit applications"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "pyyaml",
    "requests",
    "pypdf",
    "beautifulsoup4",
    "lxml",
    "rich",
    "typer",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["eval"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
